Changed alien initialization
Changed Alien interface to include init() method. New convention is that
alien constructors have no parameters, and are not supposed to do any
work. Work that could fail should happen in init(). init() also received
a context interface (api) that the alien holds on to, so we don't have
to pass it in subsequent Alien calls.

As part of this, had to change AlienContainer so that it is the one that
cause the dynamic construction of the Alien. if receives a constructor
from the SpaceGrid, along with diagnostic info like package name and
class name.